+++ example/lib/screens/api_only/input_formatters/card_number_input_formatter.dart
@@ -1,0 +1,33 @@
+import 'package:flutter/services.dart';
+
+class CardNumberInputFormatter extends TextInputFormatter {
+  static const int _groupSize = 4;
+  static const String _separator = ' ';
+
+  @override
+  TextEditingValue formatEditUpdate(
+    TextEditingValue oldValue,
+    TextEditingValue newValue,
+  ) {
+    final String newText = newValue.text;
+
+    if (newValue.selection.baseOffset == 0) {
+      return newValue;
+    }
+
+    var buffer = StringBuffer();
+    for (int i = 0; i < newText.length; i++) {
+      buffer.write(newText[i]);
+      var nonZeroIndex = i + 1;
+      if (nonZeroIndex % _groupSize == 0 && nonZeroIndex != newText.length) {
+        buffer.write(_separator);
+      }
+    }
+
+    var string = buffer.toString();
+    return newValue.copyWith(
+      text: string,
+      selection: TextSelection.collapsed(offset: string.length),
+    );
+  }
+}